Move things around, make merge words ops, etc
