+++ src/renderer/editor/clipboard.ts
