+++ src/renderer/editor/mergeSplit.ts
@@ -1,0 +1,10 @@
+import { makeMergeWords } from "renderer/store/undoStack/ops";
+import store from '../store/store';
+
+const { dispatch } = store;
+
+const mergeWords: () => void = {
+    const ranges = getSelectionRanges();
+
+    dispatchOp(makeMergeWords());
+}
